requirements.txt
pandas
openpyxl
Flask-Assets
Flask-Migrate
python-dotenv